algo_py/requirements.txt
######Algo_Ai####

# Core dependencies
numpy>=1.22,<1.24  # Adjusted to be compatible with tensorflow-intel
pandas~=1.5.0
streamlit~=1.28.0
plotly~=5.18.0
yfinance~=0.2.29
scikit-learn~=1.6.0
scipy~=1.11.0

# Cryptography and Security
pycryptodome~=3.0.0  # Modern replacement for PyCrypto
cryptography~=43.0.1
pyotp~=2.9.0

# Technical Analysis
ta~=0.10.0
pandas-ta==0.3.14b0
mplfinance~=0.12.7a17
tradingview-ta~=3.0.0
backtrader~=1.9.78
lightweight-charts==2.1.0

# Trading APIs and Tools
smartapi-python~=1.3.5
python-dotenv~=1.0.0
websocket-client~=1.6.0
nsepy==0.8  # Changed from >=1.3.0 to 0.8 (latest stable version)
nsepython
quantlib

# Web and Data Processing
logzero~=1.7.0
requests~=2.32.3
brotli~=1.1.0
beautifulsoup4~=4.12.3
simplejson~=3.19.1
filelock~=3.16.1
Flask~=3.1.0

# Visualization
matplotlib~=3.10.0
fpdf2~=2.7.0
pillow>=7.1.0,<11.0  # Updated to be compatible with streamlit

# Additional Dependencies
six~=1.16.0
python-dateutil~=2.9.0
pytz~=2024.2
certifi~=2024.2.0
charset-normalizer~=3.4.1
click~=8.1.8
colorama~=0.4.6
Jinja2~=3.1.5
joblib~=1.4.2
attrs~=23.1.0
idna~=3.4

# Machine Learning
tensorflow~=2.12.0  # Added tensorflow
xgboost

